+++ feature/picker-entity/src/main/java/com/krossovochkin/fiberyunofficial/entitypicker/domain/GetEntityTypeSchemaInteractor.kt
@@ -20,16 +20,11 @@ import com.krossovochkin.fiberyunofficial.api.FiberyApiRepository
 import com.krossovochkin.fiberyunofficial.domain.FiberyEntityTypeSchema
 import com.krossovochkin.fiberyunofficial.domain.FiberyFieldSchema
 
-interface GetEntityTypeSchemaInteractor {
-
-    suspend fun execute(fieldSchema: FiberyFieldSchema): FiberyEntityTypeSchema
-}
-
-class GetEntityTypeSchemaInteractorImpl(
+class GetEntityTypeSchemaInteractor(
     private val fiberyApiRepository: FiberyApiRepository
-) : GetEntityTypeSchemaInteractor {
+) {
 
-    override suspend fun execute(fieldSchema: FiberyFieldSchema): FiberyEntityTypeSchema {
+    suspend fun execute(fieldSchema: FiberyFieldSchema): FiberyEntityTypeSchema {
         return fiberyApiRepository.getTypeSchema(fieldSchema.type)
     }
 }
